services/gmail_api.py
import os
from dotenv import load_dotenv

load_dotenv()

class GmailService:
    def __init__(self):
        self.credentials_path = os.getenv("GMAIL_CREDENTIALS")
        # Initialize Gmail API client here if needed

    def send_email(self, to, subject, body):
        # Placeholder for sending an email
        print(f"[GmailService] Sending email to {to} with subject '{subject}' and body '{body}'")
        return "Email sent."

    def get_emails(self, query=None):
        # Placeholder for fetching emails
        print(f"[GmailService] Fetching emails with query: {query}")
        return []

    def delete_email(self, email_id):
        # Placeholder for deleting an email
        print(f"[GmailService] Deleting email: {email_id}")
        return "Email deleted."